Graphrag folder refactor.
