+++ api/settings.py
@@ -21,7 +21,7 @@ import rag.utils.infinity_conn
 
 import rag.utils
 from rag.nlp import search
-from graphrag import search as kg_search
+from graphrag.general import search as kg_search
 from api.utils import get_base_config, decrypt_database_config
 from api.constants import RAG_FLOW_SERVICE_NAME
