+++ graphrag/general/community_report_prompt.py
