+++ graphrag/general/entity_embedding.py
@@ -9,7 +9,7 @@ from typing import Any
 import numpy as np
 import networkx as nx
 from dataclasses import dataclass
-from graphrag.leiden import stable_largest_connected_component
+from graphrag.general.leiden import stable_largest_connected_component
 import graspologic as gc
