+++ graphrag/general/mind_map_extractor.py
@@ -23,8 +23,8 @@ from typing import Any
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass
 
-from graphrag.extractor import Extractor
-from graphrag.mind_map_prompt import MIND_MAP_EXTRACTION_PROMPT
+from graphrag.general.extractor import Extractor
+from graphrag.general.mind_map_prompt import MIND_MAP_EXTRACTION_PROMPT
 from graphrag.utils import ErrorHandlerFn, perform_variable_replacements
 from rag.llm.chat_model import Base as CompletionLLM
 import markdown_to_json
