+++ graphrag/general/mind_map_prompt.py
