--- lightrag/utils.py
@@ -1,66 +1,0 @@
-# Licensed under the MIT License
-"""
-Reference:
- - [LightRag](https://github.com/HKUDS/LightRAG)
-"""
-import html
-import re
-from hashlib import md5
-from typing import Any
-
-
-def compute_args_hash(*args):
-    return md5(str(args).encode()).hexdigest()
-
-
-# Refer the utils functions of the official GraphRAG implementation:
-# https://github.com/microsoft/graphrag
-def clean_str(input: Any) -> str:
-    """Clean an input string by removing HTML escapes, control characters, and other unwanted characters."""
-    # If we get non-string input, just give it back
-    if not isinstance(input, str):
-        return input
-
-    result = html.unescape(input.strip())
-    # https://stackoverflow.com/questions/4324790/removing-control-characters-from-a-string-in-python
-    return re.sub(r"[\x00-\x1f\x7f-\x9f]", "", result)
-
-
-def handle_single_entity_extraction(
-    record_attributes: list[str],
-    chunk_key: str,
-):
-    if len(record_attributes) < 4 or record_attributes[0] != '"entity"':
-        return None
-    # add this record as a node in the G
-    entity_name = clean_str(record_attributes[1].upper())
-    if not entity_name.strip():
-        return None
-    entity_type = clean_str(record_attributes[2].upper())
-    entity_description = clean_str(record_attributes[3])
-    entity_source_id = chunk_key
-    return dict(
-        entity_name=entity_name,
-        entity_type=entity_type,
-        description=entity_description,
-        source_id=entity_source_id,
-    )
-
-
-def pack_user_ass_to_openai_messages(*args: str):
-    roles = ["user", "assistant"]
-    return [
-        {"role": roles[i % 2], "content": content} for i, content in enumerate(args)
-    ]
-
-
-def split_string_by_multi_markers(content: str, markers: list[str]) -> list[str]:
-    """Split a string by multiple markers"""
-    if not markers:
-        return [content]
-    results = re.split("|".join(re.escape(marker) for marker in markers), content)
-    return [r.strip() for r in results if r.strip()]
-
-
-def is_float_regex(value):
-    return bool(re.match(r"^[-+]?[0-9]*\.?[0-9]+$", value))
