+++ rag/utils/es_conn.py
@@ -348,9 +348,11 @@ class ESConnection(DocStoreConnection):
             if (not isinstance(k, str) or not v) and k != "available_int":
                 continue
             if isinstance(v, str):
-                scripts.append(f"ctx._source.{k} = '{v}'")
+                scripts.append(f"ctx._source.{k} = '{v}';")
             elif isinstance(v, int):
-                scripts.append(f"ctx._source.{k} = {v}")
+                scripts.append(f"ctx._source.{k} = {v};")
+            elif isinstance(v, list):
+                scripts.append(f"ctx._source.{k} = {json.dumps(v)};")
             else:
                 raise Exception(
                     f"newValue `{str(k)}={str(v)}` value type is {str(type(v))}, expected to be int, str.")
